+++ functions/.husky-prepare.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+# Husky pre-commit hook preparation script
+npm run precommit